backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.11.0
PyJWT==2.10.1
openai==2.29.0
httpx==0.28.1
qdrant-client==1.15.1
sqlparse==0.5.3
python-dotenv==1.1.1
python-multipart==0.0.20
openpyxl==3.1.5
sentence-transformers==5.1.2
pytest==8.4.2
